updated almost everything
